requirements.txt
pandas>=1.5
numpy>=1.24
pyyaml
lightgbm
scikit-learn
joblib
matplotlib
gymnasium
ccxt
requests
fastapi
uvicorn
python-dotenv
schedule
redis
aioredis
websockets
asyncio
ta
plotly
seaborn
jupyter
pytest
